remove qt4 version of qtsingleapplication, take significant updates of qt5 version from orig repo
