+++ app/src/libraries/qtSingleApplication/qtsingleapplication.pri
@@ -2,6 +2,7 @@ include(../common.pri)
 INCLUDEPATH += $$PWD
 DEPENDPATH += $$PWD
 QT *= network
+greaterThan(QT_MAJOR_VERSION, 4): QT *= widgets
 
 qtsingleapplication-uselib:!qtsingleapplication-buildlib {
     LIBS += -L$$QTSINGLEAPPLICATION_LIBDIR -l$$QTSINGLEAPPLICATION_LIBNAME
