--- app/src/libraries/qtSingleApplication5/QtLockedFile
@@ -1,1 +1,0 @@
-#include "qtlockedfile.h"
