--- app/src/libraries/qtSingleApplication5/QtSingleApplication
@@ -1,1 +1,0 @@
-#include "qtsingleapplication.h"
